Bump sdk version, move tests to more appropriate place
