+++ firely-validator-api-tests.props
@@ -8,7 +8,7 @@
 	</PropertyGroup>
 
 	<PropertyGroup>
-		<FirelySdkVersion>5.11.4</FirelySdkVersion>
+		<FirelySdkVersion>5.11.5-20250501.6</FirelySdkVersion>
 	</PropertyGroup>
 
 	<ItemGroup>
